I think I'm gonna need this pattern for getting out of the advanceToTime quagmire
This commit was sponsored by Al Sweigart, Brian Grohe, Greg Back, and
my other patrons.  If you want to join them, you can support my work
at https://glyph.im/patrons/.
